flask-backend/requirements.txt
deepforest==1.3.3
Flask==3.0.2
flask_cors==4.0.0
flask_sqlalchemy==3.1.1
lightning==2.2.1
matplotlib==3.5.2
numpy==1.22.4
Pillow==9.1.1
Pillow==10.2.0
pytorch_lightning==2.1.2
torch==2.1.1
torchmetrics==1.2.0
torchvision==0.16.1
